fix: Possible panic
If one of the goroutines spawned in listBigqueryDatasetTables fails and
returns an error then errgroup.Group.Go will cancel the context.

Then semaphore.Acquire can fail because of the canceled context. After this resolver
finishes, and the SDK closes results channel. But other spawned goroutines can still
be running and they may try to send through this closed channel. Panic.